init; copy yaml and python script from ompi
These two files originated from the OMPI repository (diverged here:
https://github.com/open-mpi/ompi/commit/ef866a4657c782fef04513ee7334c7b762783b5e)

An OMPI maintainer wanted this and the label bot in two of their other repos.
Instead of copying the code to their repo(s), I think it'd would be far simpler
to move the actions to separate repos and publish them on the GitHub marketplace.